+++ autoware_utils/README.md
@@ -33,20 +33,10 @@ The geometry module provides classes and functions for handling 2D and 3D points
 
 The ROS module provides utilities for working with ROS messages and nodes:
 
-- **`debug_publisher.hpp`**: A helper class for publishing debug messages with timestamps.
 - **`diagnostics_interface.hpp`**: An interface for publishing diagnostic messages.
 - **`msg_covariance.hpp`**: Indices for accessing covariance matrices in ROS messages.
 - **`msg_operation.hpp`**: Overloaded operators for quaternion messages.
-- **`processing_time_publisher.hpp`**: Publishes processing times as diagnostic messages.
-- **`published_time_publisher.hpp`**: Tracks and publishes the time when messages are published.
 - **`self_pose_listener.hpp`**: Listens to the self-pose of the vehicle.
-- **`debug_traits.hpp`**: Traits for identifying debug message types.
-
-#### System Module
-
-The system module provides low-level utilities for performance monitoring and error handling:
-
-- **`time_keeper.hpp`**: Tracks and reports the processing time of various functions.
 
 ## Usage
 
@@ -91,33 +81,6 @@ int main() {
 
 ### Detailed Usage Examples
 
-#### Logging Processing Times with ProcessingTimePublisher
-
-```cpp
-#include "autoware_utils/ros/processing_time_publisher.hpp"
-#include <rclcpp/rclcpp.hpp>
-#include <map>
-
-int main(int argc, char * argv[]) {
-  rclcpp::init(argc, argv);
-  auto node = rclcpp::Node::make_shared("processing_time_node");
-
-  // Initialize ProcessingTimePublisher
-  autoware_utils::ProcessingTimePublisher processing_time_pub(node.get(), "~/debug/processing_time_ms");
-
-  // Simulate some processing times
-  std::map<std::string, double> processing_times = {
-    {"node1", 0.1}, {"node2", 0.2}, {"node3", 0.3}
-  };
-
-  // Publish processing times
-  processing_time_pub.publish(processing_times);
-
-  rclcpp::shutdown();
-  return 0;
-}
-```
-
 #### Manipulating Polygons with boost_polygon_utils.hpp
 
 ```cpp
@@ -141,29 +104,6 @@ int main(int argc, char * argv[]) {
 
   // Check if the polygon is clockwise
   bool is_clockwise = autoware_utils::is_clockwise(polygon);
-
-  rclcpp::shutdown();
-  return 0;
-}
-```
-
-#### Handling Debug Message Types with debug_traits.hpp
-
-```cpp
-#include "autoware_utils/ros/debug_publisher.hpp"
-#include "autoware_utils/ros/debug_traits.hpp"
-#include <rclcpp/rclcpp.hpp>
-
-int main(int argc, char * argv[]) {
-  rclcpp::init(argc, argv);
-  auto node = rclcpp::Node::make_shared("debug_node");
-
-  // Initialize DebugPublisher
-  autoware_utils::DebugPublisher debug_pub(node, "/debug");
-
-  // Publish a debug message with custom type
-  float debug_data = 42.0;
-  debug_pub.publish<autoware_internal_debug_msgs::msg::Float32Stamped>("example", debug_data);
 
   rclcpp::shutdown();
   return 0;
