+++ autoware_utils/include/autoware_utils/ros/debug_traits.hpp
@@ -1,4 +1,4 @@
-// Copyright 2020 Tier IV, Inc.
+// Copyright 2025 The Autoware Contributors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -15,77 +15,14 @@
 #ifndef AUTOWARE_UTILS__ROS__DEBUG_TRAITS_HPP_
 #define AUTOWARE_UTILS__ROS__DEBUG_TRAITS_HPP_
 
-#include <autoware_internal_debug_msgs/msg/bool_stamped.hpp>
-#include <autoware_internal_debug_msgs/msg/float32_multi_array_stamped.hpp>
-#include <autoware_internal_debug_msgs/msg/float32_stamped.hpp>
-#include <autoware_internal_debug_msgs/msg/float64_multi_array_stamped.hpp>
-#include <autoware_internal_debug_msgs/msg/float64_stamped.hpp>
-#include <autoware_internal_debug_msgs/msg/int32_multi_array_stamped.hpp>
-#include <autoware_internal_debug_msgs/msg/int32_stamped.hpp>
-#include <autoware_internal_debug_msgs/msg/int64_multi_array_stamped.hpp>
-#include <autoware_internal_debug_msgs/msg/int64_stamped.hpp>
-#include <autoware_internal_debug_msgs/msg/string_stamped.hpp>
+// NOLINTBEGIN(build/namespaces, whitespace/line_length)
+// clang-format off
 
-#include <type_traits>
+#pragma message("#include <autoware_utils/ros/debug_traits.hpp> is deprecated. Use #include <autoware_utils_debug/debug_traits.hpp> instead.")
+#include <autoware_utils_debug/debug_traits.hpp>
+namespace autoware_utils { using namespace autoware_utils_debug; }
 
-namespace autoware_utils::debug_traits
-{
-template <typename T>
-struct is_debug_message : std::false_type
-{
-};
-
-template <>
-struct is_debug_message<autoware_internal_debug_msgs::msg::BoolStamped> : std::true_type
-{
-};
-
-template <>
-struct is_debug_message<autoware_internal_debug_msgs::msg::Float32MultiArrayStamped>
-: std::true_type
-{
-};
-
-template <>
-struct is_debug_message<autoware_internal_debug_msgs::msg::Float32Stamped> : std::true_type
-{
-};
-
-template <>
-struct is_debug_message<autoware_internal_debug_msgs::msg::Float64MultiArrayStamped>
-: std::true_type
-{
-};
-
-template <>
-struct is_debug_message<autoware_internal_debug_msgs::msg::Float64Stamped> : std::true_type
-{
-};
-
-template <>
-struct is_debug_message<autoware_internal_debug_msgs::msg::Int32MultiArrayStamped> : std::true_type
-{
-};
-
-template <>
-struct is_debug_message<autoware_internal_debug_msgs::msg::Int32Stamped> : std::true_type
-{
-};
-
-template <>
-struct is_debug_message<autoware_internal_debug_msgs::msg::Int64MultiArrayStamped> : std::true_type
-{
-};
-
-template <>
-struct is_debug_message<autoware_internal_debug_msgs::msg::Int64Stamped> : std::true_type
-{
-};
-
-template <>
-struct is_debug_message<autoware_internal_debug_msgs::msg::StringStamped> : std::true_type
-{
-};
-}  // namespace autoware_utils::debug_traits
+// clang-format on
+// NOLINTEND
 
 #endif  // AUTOWARE_UTILS__ROS__DEBUG_TRAITS_HPP_
