+++ autoware_utils_debug/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>autoware_utils</name>
-  <version>1.2.0</version>
-  <description>The autoware_utils package</description>
+  <name>autoware_utils_debug</name>
+  <version>1.1.0</version>
+  <description>The autoware_utils_debug package</description>
   <maintainer email="egon.kang@autocore.ai">Jian Kang</maintainer>
   <maintainer email="ryohsuke.mitsudome@tier4.jp">Ryohsuke Mitsudome</maintainer>
   <maintainer email="esteve.fernandez@tier4.jp">Esteve Fernandez</maintainer>
@@ -16,28 +16,9 @@
 
   <depend>autoware_internal_debug_msgs</depend>
   <depend>autoware_internal_msgs</depend>
-  <depend>autoware_internal_planning_msgs</depend>
-  <depend>autoware_perception_msgs</depend>
-  <depend>autoware_planning_msgs</depend>
-  <depend>autoware_utils_debug</depend>
-  <depend>autoware_utils_geometry</depend>
-  <depend>autoware_utils_logging</depend>
-  <depend>autoware_utils_math</depend>
-  <depend>autoware_utils_pcl</depend>
-  <depend>autoware_utils_rclcpp</depend>
   <depend>autoware_utils_system</depend>
-  <depend>autoware_utils_uuid</depend>
-  <depend>autoware_utils_visualization</depend>
-  <depend>autoware_vehicle_msgs</depend>
-  <depend>builtin_interfaces</depend>
   <depend>diagnostic_msgs</depend>
-  <depend>geometry_msgs</depend>
-  <depend>libboost-system-dev</depend>
   <depend>rclcpp</depend>
-  <depend>tf2</depend>
-  <depend>tf2_eigen</depend>
-  <depend>tf2_geometry_msgs</depend>
-  <depend>unique_identifier_msgs</depend>
 
   <test_depend>ament_cmake_ros</test_depend>
   <test_depend>ament_lint_auto</test_depend>
