+++ autoware_utils_debug/test/cases/debug_publisher.cpp
@@ -12,17 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE_UTILS__ROS__DEBUG_PUBLISHER_HPP_
-#define AUTOWARE_UTILS__ROS__DEBUG_PUBLISHER_HPP_
+#include "autoware_utils_debug/debug_publisher.hpp"
 
-// NOLINTBEGIN(build/namespaces, whitespace/line_length)
-// clang-format off
+#include <gtest/gtest.h>
 
-#pragma message("#include <autoware_utils/ros/debug_publisher.hpp> is deprecated. Use #include <autoware_utils_debug/debug_publisher.hpp> instead.")
-#include <autoware_utils_debug/debug_publisher.hpp>
-namespace autoware_utils { using namespace autoware_utils_debug; }
+#include <memory>
 
-// clang-format on
-// NOLINTEND
+using autoware_utils_debug::DebugPublisher;
 
-#endif  // AUTOWARE_UTILS__ROS__DEBUG_PUBLISHER_HPP_
+TEST(TestDebugPublisher, Instantiation)
+{
+  const auto node = std::make_shared<rclcpp::Node>("test_node");
+  DebugPublisher(node.get(), "namespace");
+}
